Edit JavaSript

+++ database/migrations/2021_01_07_000021_create_content_content_category_pivot_table.php
@@ -9,9 +9,9 @@ class CreateContentContentCategoryPivotTable extends Migration
     public function up()
     {
         Schema::create('content_content_category', function (Blueprint $table) {
-            $table->unsignedInteger('content_id');
+            $table->unsignedBigInteger('content_id');
             $table->foreign('content_id', 'content_id_fk_2912306')->references('id')->on('contents')->onDelete('cascade');
-            $table->unsignedInteger('content_category_id');
+            $table->unsignedBigInteger('content_category_id');
             $table->foreign('content_category_id', 'content_category_id_fk_2912306')->references('id')->on('content_categories')->onDelete('cascade');
         });
     }
