+++ database/migrations/2021_01_07_000022_create_class_room_school_pivot_table.php
@@ -9,9 +9,9 @@ class CreateClassRoomSchoolPivotTable extends Migration
     public function up()
     {
         Schema::create('class_room_school', function (Blueprint $table) {
-            $table->unsignedInteger('class_room_id');
+            $table->unsignedBigInteger('class_room_id');
             $table->foreign('class_room_id', 'class_room_id_fk_2911205')->references('id')->on('class_rooms')->onDelete('cascade');
-            $table->unsignedInteger('school_id');
+            $table->unsignedBigInteger('school_id');
             $table->foreign('school_id', 'school_id_fk_2911205')->references('id')->on('schools')->onDelete('cascade');
         });
     }
